Move registration components to new subdirectory
